refactor(intents): extract view methods from helpers.go to views.go
Reduce helpers.go from 552 to 345 lines by extracting view methods:

- wizardView, viewExportComplete, exportFormatDisplayName
- getTheme, getCardStyle, getWizardBreadcrumbs, getWizardContextHelp
- renderWizardModalOverlay, renderProgressModalOverlay, renderExportModalOverlay
- renderReviewScreenWithModalOverlay, renderPreviewScreenWithModalOverlay

New views.go file: 213 lines
helpers.go now under 500 line max (was 552, now 345)

All 164 generatecv tests pass.

AI-Generated-By: OpenCode (Claude Sonnet 4)
